refactor LoadErrorModal to interact with store itself rather than via props
